Now a header only library
