--- src/dcgp.h
@@ -1,9 +1,0 @@
-#ifndef DCGP_H
-#define DCGP_H
-
-#include "expression.h"
-#include "fitness_functions.h"
-#include "function_set.h"
-#include "std_overloads.h"
-
-#endif // DCGP_H
