+++ src/dcgp.hpp
@@ -1,0 +1,7 @@
+#ifndef DCGP_H
+#define DCGP_H
+
+#include "expression.hpp"
+#include "function_set.hpp"
+
+#endif // DCGP_H
