+++ src/wrapped_functions.hpp
