+++ tests/CMakeLists.txt
@@ -6,7 +6,7 @@ ADD_DEFINITIONS(-DBOOST_TEST_DYN_LINK)
 MACRO(ADD_DCGP_TESTCASE arg1)
 	IF(CMAKE_BUILD_TYPE STREQUAL "Debug")
 		ADD_EXECUTABLE(${arg1} ${arg1}.cpp)
-		TARGET_LINK_LIBRARIES(${arg1} dcgp_s ${MANDATORY_LIBRARIES} ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY})
+		TARGET_LINK_LIBRARIES(${arg1} ${MANDATORY_LIBRARIES} ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY})
 		ADD_TEST(${arg1} ${arg1})
 	ENDIF(CMAKE_BUILD_TYPE STREQUAL "Debug")
 ENDMACRO(ADD_DCGP_TESTCASE)
@@ -14,7 +14,7 @@ ENDMACRO(ADD_DCGP_TESTCASE)
 MACRO(ADD_DCGP_PERFORMANCE_TESTCASE arg1)
 	IF(CMAKE_BUILD_TYPE STREQUAL "Release")
 		ADD_EXECUTABLE("${arg1}_perf" "${arg1}_perf.cpp")
-		TARGET_LINK_LIBRARIES("${arg1}_perf" dcgp_s ${MANDATORY_LIBRARIES} ${Boost_CHRONO_LIBRARY} ${Boost_TIMER_LIBRARY} ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY} ${Boost_SYSTEM_LIBRARY})
+		TARGET_LINK_LIBRARIES("${arg1}_perf" ${MANDATORY_LIBRARIES} ${Boost_CHRONO_LIBRARY} ${Boost_TIMER_LIBRARY} ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY} ${Boost_SYSTEM_LIBRARY})
 		ADD_TEST("${arg1}_perf" "${arg1}_perf")
 	ENDIF(CMAKE_BUILD_TYPE STREQUAL "Release")
 ENDMACRO(ADD_DCGP_PERFORMANCE_TESTCASE)
