+++ tests/evolving_simple_expressions.cpp
@@ -3,7 +3,7 @@
 #include <random>
 #include <cmath>
 
-#include "../src/dcgp.h"
+#include "../src/dcgp.hpp"
 
 bool test_fails(
         unsigned int r,
